+++ backend/canisters/community/impl/src/updates/send_message.rs
@@ -17,9 +17,10 @@ use community_canister::c2c_send_message::{Args as C2CArgs, Response as C2CRespo
 use community_canister::send_message::{Response::*, *};
 use constants::{MEMO_MESSAGE, MEMO_PRIZE};
 use group_chat_core::SendMessageSuccess;
-use group_community_common::{MemberTransfer, NewP2PSwap, prize_refund, validate_prize};
+use group_community_common::{NewP2PSwap, prize_refund, validate_prize};
 use itertools::Itertools;
 use lazy_static::lazy_static;
+use ledger_utils::UserTransfer;
 use oc_error_codes::OCErrorCode;
 use regex_lite::Regex;
 use std::str::FromStr;
@@ -664,7 +665,7 @@ fn prepare_transfer(args: &Args, state: &mut RuntimeState) -> OCResult<(UserId, 
             }
         };
 
-    let transfer = MemberTransfer::new(transfer, recipient, memo, this_canister_id)?;
+    let transfer = UserTransfer::new(transfer, recipient, memo, this_canister_id)?;
 
     state
         .data
@@ -680,8 +681,8 @@ fn prepare_transfer(args: &Args, state: &mut RuntimeState) -> OCResult<(UserId, 
         )?;
 
     match transfer {
-        MemberTransfer::Icrc2(transfer) => Ok((user_id, PrepareTransferResult::Icrc2(transfer))),
-        MemberTransfer::Certified(transfer) => {
+        UserTransfer::Icrc2(transfer) => Ok((user_id, PrepareTransferResult::Icrc2(transfer))),
+        UserTransfer::Certified(transfer) => {
             let completed = state.data.certified_transfers.verify(
                 transfer,
                 state.env.caller(),